+++ tests/slr_integration/Admin/Ajax_Test.php
@@ -62,94 +62,6 @@ class Ajax_Test extends Controller_Test_Case {
 		Sessions::truncate();
 	}
 
-	/**
-	 * @test
-	 */
-	public function it_should_filter_seating_totals(): void {
-		$post_id = static::factory()->post->create();
-
-		$tickets_handler = tribe( 'tickets.handler' );
-
-		update_post_meta( $post_id, Meta::META_KEY_LAYOUT_ID, 'layout-uuid-1' );
-		update_post_meta( $post_id, Meta::META_KEY_ENABLED, '1' );
-		update_post_meta( $post_id, Global_Stock::GLOBAL_STOCK_ENABLED, '1' );
-		update_post_meta( $post_id, Global_Stock::GLOBAL_STOCK_LEVEL, 40 );
-		update_post_meta( $post_id, $tickets_handler->key_capacity, 40 );
-
-		$ticket_id_1 = $this->create_tc_ticket(
-			$post_id,
-			10,
-			[
-				'tribe-ticket' => [
-					'mode'     => Global_Stock::CAPPED_STOCK_MODE,
-					'capacity' => 40,
-				],
-			]
-		);
-
-		$ticket_id_2 = $this->create_tc_ticket(
-			$post_id,
-			10,
-			[
-				'tribe-ticket' => [
-					'mode'     => Global_Stock::CAPPED_STOCK_MODE,
-					'capacity' => 40,
-				],
-			]
-		);
-
-		$ticket_id_3 = $this->create_tc_ticket(
-			$post_id,
-			10,
-			[
-				'tribe-ticket' => [
-					'mode'     => Global_Stock::CAPPED_STOCK_MODE,
-					'capacity' => 40,
-				],
-			]
-		);
-
-		update_post_meta( $ticket_id_1, Meta::META_KEY_ENABLED, true );
-		update_post_meta( $ticket_id_1, Meta::META_KEY_SEAT_TYPE, 'some-seat-type-1' );
-
-		update_post_meta( $ticket_id_2, Meta::META_KEY_ENABLED, true );
-		update_post_meta( $ticket_id_2, Meta::META_KEY_SEAT_TYPE, 'some-seat-type-1' );
-
-		update_post_meta( $ticket_id_3, Meta::META_KEY_ENABLED, true );
-		update_post_meta( $ticket_id_3, Meta::META_KEY_SEAT_TYPE, 'some-seat-type-1' );
-
-		$order = $this->create_order(
-			[
-				$ticket_id_1 => 3,
-				$ticket_id_2 => 5,
-			]
-		);
-
-		$totals_1 = [
-			'stock'   => 40 - 3 - 5,
-			'sold'    => 3,
-			'pending' => 0,
-		];
-
-		$totals_2 = [
-			'stock'   => 40 - 3 - 5,
-			'sold'    => 5,
-			'pending' => 0,
-		];
-
-		$totals_3 = [
-			'stock'   => 40 - 3 - 5,
-			'sold'    => 0,
-			'pending' => 0,
-		];
-
-		$this->make_controller()->register();
-
-		$this->assertEquals( $totals_1, $tickets_handler->get_ticket_totals( get_post( $ticket_id_1 ) ) );
-		$this->assertEquals( $totals_2, $tickets_handler->get_ticket_totals( get_post( $ticket_id_2 ) ) );
-		$this->assertEquals( $totals_3, $tickets_handler->get_ticket_totals( get_post( $ticket_id_3 ) ) );
-	}
-
 	public function test_get_localized_data(): void {
 		$this->set_fn_return( 'wp_create_nonce', '88b1a4b166' );
 		$controller = $this->make_controller();
